refactor(WhatsNewDialog): consolidate files
