+++ client/src/dialogs/DialogRoot.jsx
@@ -10,7 +10,7 @@ import GeotagDialog from './Geotag'
 import SaveAsImageDialog from './SaveAsImage'
 import SettingsDialog from './Settings'
 import SignInDialog from './SignInDialog'
-import WhatsNewDialog from './WhatsNewDialog'
+import WhatsNewDialog from './WhatsNew'
 import NewsletterDialog from './Newsletter'
 import UpgradeDialog from './UpgradeDialog'
 import SentimentSurveyDialog from './SentimentSurveyDialog'
