+++ client/src/dialogs/WhatsNew/WhatsNewDialog.scss
