+++ client/src/dialogs/WhatsNew/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WhatsNewDialog'
